fix: PodMonitor
* crd-replicator

Commit 10ef3087 changed the default port for metrics from 8080 to 8082,
but crd-replicator was not changed because there is no command line option.
However, only the deployment port was changed to 8082, so PodMonitor no
longer works properly.

This commit adds the command line option, which defaults to 8082.

* virtualkubelet

Commit ddafea075 changed the label of VirtualKubelet but did not change
the PodMonitor selector so it no longer works correctly.

This commit makes PodMonitor's selector follow commit ddafea075.

Signed-off-by: Mitsuru Kariya <mitsuru.kariya@nttdata.com>